CMakeLists Update, gfortran compatibility update
CMakeLists file now provides the command to use ifort by default when building. Can be changed to gfortran if you desire. Found an issue with getcwd() from intrinsic library, it was set here to output the string to an unallocated dynamic array, and this worked fine when compiled with ifort. gfortran does not like this, however, and requires a properly allocated string. Have just gone for a long ass string of 300 characters and trimmed it, might fix this later by using a long buffer and writing string to an appropriately sized string after.